add: refactor code + add form post action
