+++ classes/Dragon.php
@@ -1,5 +1,11 @@
 <?php
 
+namespace app\classes;
+
+/**
+ * Class Dragon
+ * @package app\classes
+ */
 class Dragon
 {
