+++ classes/Engine.php
@@ -2,78 +2,62 @@
 
 namespace app\classes;
 
+require_once('InterfaceApp.php');
+
+/**
+ * Class Engine
+ * @package app\classes
+ */
+
 class Engine
 {
-    private $output;
-    private $number_level_map = 1;
+    /**
+     * @var array
+     */
     private $errors = [];
 
+    /**
+     * @return array
+     */
     public function show_error()
     {
         return $this->errors;
     }
 
-    private function renderInterface($map)
-    {
-        $this->output .= '<div class="main-interface">';
-
-        $this->output .= $this->renderFormInterface();
-
-        $this->output .= $map;
-
-        $this->output .= '</div>';
+    private function initMap($map){
+        $map->loadLevelOnTheMap();
+        $map->renderMap();
     }
 
-    private function renderFormInterface()
-    {
-        return '
-            <div class="nav-state">
-                <div class="nav-status">
-                    <p>Level Game is  '."$this->number_level_map".'</p>
-                    <button class="btn btn-danger" value=\'document.cookie ="php_game_level_number=1; expires = Thu, 01 Jan 1970 00:00:00 GMT"\'>Reset Game</button>
-                </div>        
-                <div class="nav-bar">
-                     <form method="POST" >
-                        <div class="nav-bar-top">
-                            <button type="button" class="btn btn-primary">Move Up</button>
-                        </div>
-                        <div class="nav-bar-center">
-                            <button type="button" class="btn btn-primary">Move Left</button>
-                            <button type="button" class="btn btn-primary">Move Right</button>
-                        </div>
-                        <div class="nav-bar-center">
-                             <button type="button" class="btn btn-primary">Move Down</button>
-                        </div>          
-                    </form>
-                </div>    
-            </div>
-        ';
-    }
-
-    private function renderPage($htmlMap)
-    {
-        $this->output = file_get_contents('./templates/header.php');
-
-        $this->renderInterface($htmlMap);
-
-        $this->output .= file_get_contents('./templates/footer.php');
-
-        return $this->output;
-    }
-
+    /**
+     *
+     */
     public function run()
     {
         global $OUTPUT;
 
         $map = new Map();
-        $map->loadLevelOnTheMap();
-        $map->renderMap();
 
-        $htmlMap = $map->getGridOfMap();
+        if($map instanceof Map) {
+            $this -> initMap($map);
+        }
 
-        $OUTPUT = $this->renderPage($htmlMap);
+        $interface = new InterfaceApp();
+
+        $player = new Player();
+
+        if($_POST['playerDirectionMove'] ?? null){
+            $player ->movePlayer();
+        }
+
+        $OUTPUT = $interface->renderPage($map->getGridOfMap());
     }
 
+    /**
+     * @param $valuePointMapToGoThePlayer
+     * @param string $finish
+     * @return bool
+     */
     private function checkIfPlayerIsFinish($valuePointMapToGoThePlayer, $finish = 'D')
     {
         if ($valuePointMapToGoThePlayer == $finish) {
