+++ classes/Player.php
@@ -1,36 +1,57 @@
 <?php
 
+namespace app\classes;
+
+/**
+ * Class Player
+ */
 class Player
 {
+    /**
+     * @var
+     */
     private $playerIsAlive;
+    /**
+     * @var
+     */
     private $playerPositionKey;
+    /**
+     * @var
+     */
     private $playerPositionKeyRow;
+    /**
+     * @var bool
+     */
     private $playerMovedToNewPointOnTheMap = false;
+    /**
+     * @var bool
+     */
     private $playerCollectedCoins = true;
+    /**
+     * @var string[]
+     */
     private $playerAvailableWaysOnTheMap = ['.', 'C'];
+    /**
+     * @var null
+     */
     private $map = null;
 
-    public function movePlayer()
-    {
-        $this->map = new Map();
-        foreach ($this->map as $keyRowMap => $rowMapArray) {
-            $this->checkIfPlayerHaveAllCoins($this->map, $keyRowMap);
-            if ($playerKeyOnTheRowMap = array_search('P', $this->map[$keyRowMap], true)) {
-                $this->definePlayerPosition($playerKeyOnTheRowMap, $keyRowMap);
-                $this->moveLeft($this->map, $rowMapArray, $_POST, $rowMapArray[$playerKeyOnTheRowMap - 1], $this->playerAvailableWaysOnTheMap);
-                $this->moveRight($this->map, $rowMapArray, $_POST, $rowMapArray[$playerKeyOnTheRowMap + 1], $this->playerAvailableWaysOnTheMap);
-                $this->moveUp($this->map, $_POST, $this->playerAvailableWaysOnTheMap);
-                $this->moveDown($this->map, $_POST, $this->playerAvailableWaysOnTheMap);
-            }
-        }
-    }
 
+    /**
+     * @param $keyPlayerInArray
+     * @param $keyPlayerRowArray
+     */
     public function definePlayerPosition($keyPlayerInArray, $keyPlayerRowArray)
     {
         $this->playerPositionKey = $keyPlayerInArray;
         $this->playerPositionKeyRow = $keyPlayerRowArray;
     }
 
+    /**
+     * @param $mapArray
+     * @param $keyRowMapArray
+     * @return bool
+     */
     public function checkIfPlayerHaveAllCoins($mapArray, $keyRowMapArray)
     {
         if ($keyCoins = array_search('C', $mapArray[$keyRowMapArray], true)) {
@@ -42,7 +63,15 @@ class Player
         }
     }
 
-    public function moveLeft($mapArray, $rowMapArray, $post, $pointPlayerWillBeOnTheMap, $arrAvailableWaysOnTheMap)
+    /**
+     * @param $mapArray
+     * @param $rowMapArray
+     * @param $post
+     * @param $pointPlayerWillBeOnTheMap
+     * @param $arrAvailableWaysOnTheMap
+     * @return false
+     */
+    private function moveLeft($mapArray, $rowMapArray, $post, $pointPlayerWillBeOnTheMap, $arrAvailableWaysOnTheMap)
     {
         if ($post['playerDirectionMove'] == 'left' and in_array($pointPlayerWillBeOnTheMap, $arrAvailableWaysOnTheMap)) {
             $mapArray[$this->playerPositionKeyRow][$this->playerPositionKey] = '.';
@@ -58,7 +87,15 @@ class Player
         }
     }
 
-    public function moveRight($mapArray, $rowMapArray, $post, $nextPointPlayerWillBe, $arrAvailableWaysOnTheMap)
+    /**
+     * @param $mapArray
+     * @param $rowMapArray
+     * @param $post
+     * @param $nextPointPlayerWillBe
+     * @param $arrAvailableWaysOnTheMap
+     * @return false
+     */
+    private function moveRight($mapArray, $rowMapArray, $post, $nextPointPlayerWillBe, $arrAvailableWaysOnTheMap)
     {
         if ($post['playerDirectionMove'] == 'right' and in_array($nextPointPlayerWillBe, $arrAvailableWaysOnTheMap)) {
             $mapArray[$this->playerPositionKeyRow][$this->playerPositionKey] = '.';
@@ -74,7 +111,12 @@ class Player
         }
     }
 
-    public function moveUp($mapArray, $post, $arrAvailableWaysOnTheMap)
+    /**
+     * @param $mapArray
+     * @param $post
+     * @param $arrAvailableWaysOnTheMap
+     */
+    private function moveUp($mapArray, $post, $arrAvailableWaysOnTheMap)
     {
         if ($this->playerMovedToNewCellOnTheMap == false) {
             if ($post['playerDirectionMove'] == 'up' and in_array($mapArray[$this->playerPositionKeyRow - 1][$this->playerPositionKey], $arrAvailableWaysOnTheMap)) {
@@ -88,7 +130,12 @@ class Player
         }
     }
 
-    public function moveDown($mapArray, $post, $arrAvailableWaysOnTheMap)
+    /**
+     * @param $mapArray
+     * @param $post
+     * @param $arrAvailableWaysOnTheMap
+     */
+    private function moveDown($mapArray, $post, $arrAvailableWaysOnTheMap)
     {
         if ($this->playerMovedToNewCellOnTheMap == false) {
             if ($post['playerDirectionMove'] == 'down' and in_array($mapArray[$this->playerPositionKeyRow + 1][$this->playerPositionKey], $arrAvailableWaysOnTheMap)) {
@@ -98,6 +145,24 @@ class Player
                 $this->playerPositionKeyRow += 1;
 
                 $this->playerMovedToNewPointOnTheMap = true;
+            }
+        }
+    }
+
+    /**
+     *
+     */
+    public function movePlayer()
+    {
+        $this->map = new Map();
+        foreach ($this->map as $keyRowMap => $rowMapArray) {
+            $this->checkIfPlayerHaveAllCoins($this->map, $keyRowMap);
+            if ($playerKeyOnTheRowMap = array_search('P', $this->map[$keyRowMap], true)) {
+                $this->definePlayerPosition($playerKeyOnTheRowMap, $keyRowMap);
+                $this->moveLeft($this->map, $rowMapArray, $_POST, $rowMapArray[$playerKeyOnTheRowMap - 1], $this->playerAvailableWaysOnTheMap);
+                $this->moveRight($this->map, $rowMapArray, $_POST, $rowMapArray[$playerKeyOnTheRowMap + 1], $this->playerAvailableWaysOnTheMap);
+                $this->moveUp($this->map, $_POST, $this->playerAvailableWaysOnTheMap);
+                $this->moveDown($this->map, $_POST, $this->playerAvailableWaysOnTheMap);
             }
         }
     }
