+++ index.php
@@ -2,6 +2,8 @@
 
 namespace app;
 
+use app\classes\Engine;
+
 require_once(__DIR__ . '/lang/install.php');
 require_once(__DIR__ . '/config.php');
 require_once(__DIR__ . '/lib.php');
@@ -10,7 +12,6 @@ require_once(__DIR__ . '/classes/Engine.php');
 require_once(__DIR__ . '/classes/Player.php');
 require_once(__DIR__ . '/classes/Map.php');
 
-use app\classes\Engine;
 
 $gameEngine = new Engine();
 $gameEngine->run();
